pyproject.toml
[tool.poetry]
name = "lectio-recommender"
version = "0.1.0"
description = "Sistema de recomendação de livros baseado em embeddings e grafos"
authors = ["Iury Amaral <iurydrayton@gmail.com>"]
readme = "README.md"
packages = [{ include = "lectio_recommender", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11"
uvicorn = ">=0.41.0,<0.42.0"
fastapi = ">=0.135.1,<0.136.0"
sqlalchemy = ">=2.0.48,<3.0.0"
python-dotenv = "^1.0.0"   # corrigido o nome oficial
psycopg2-binary = ">=2.9.11,<3.0.0"
pydantic = ">=2.12.5,<3.0.0"
faker = ">=40.11.0,<41.0.0"
neo4j = ">=6.1.0,<7.0.0"

[tool.poetry.group.trainer.dependencies]
tensorflow = ">=2.16,<2.21.0 || >2.21.0"
numpy = ">=1.26,<2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]   # não precisa fixar <3.0 se não houver motivo
build-backend = "poetry.core.masonry.api"